third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-row-block-end-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "10"
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "-10"
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "63"
Harness: the test ran to completion.
